+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to `homebridge` will be documented in this file. This projec
 
 - chore: update `actions/checkout` to `v7`
 - chore: dependency updates
+- fix: keep api shutdown listener attached across child bridge restart
 
 ### Homebridge Dependencies